+++ NEWS.md
@@ -1,5 +1,7 @@
 # dtplyr (development version)
 
+* `if_any()` and `if_all()` now work without specifying the `.fns` argument (@mgirlich, #325).
+
 # dtplyr 1.2.1
 
 * Fix for upcoming rlang release.